Solution to mismatched tensor dimension issue.
Mismatched tensor dimension error thrown when attempting to calculate expected_state_action_values for batch. This solution uses the expand_by() function to resize the reward_batch tensor to the size of next_state_values. This is fine because reward_batch is a constant tensor (each element is the same scalar reward value).